requirements.txt
numpy
matplotlib
pyyaml
pygame
scipy
collision
corgipath